requirements.txt
numpy==1.26.4
Pillow==10.3.0
scikit_learn==1.4.2
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
porespy==2.4.1
